services/reconciliation-worker/requirements.txt
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pydantic==2.5.3
python-dotenv==1.0.0
schedule==1.2.1
requests==2.31.0
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-instrumentation-sqlalchemy==0.43b0
